web/requirements.txt
Flask==1.0.2
Flask-Markdown==0.3
Flask-PageDown==0.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.3
Jinja2==2.10
MarkupSafe==1.1.1
SQLAlchemy==1.3.1
Werkzeug==0.15.1
gunicorn==19.9.0
python-slugify==4.0.1
itsdangerous==1.1.0
psycopg2==2.7.7